+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Switching .sav would not take effect until a ROM was loaded.
 
 ### Added
- - Kit Editor: "Reload samples", "Save ROM" buttons.
+ - Kit Editor: "Reload samples", "Save ROM", "Clear kit" buttons.
  - Kit Editor: .lsdkit files now include original 16-bit sample data as well as local file path, to make future edits easier.
  - Kit Editor: automatic silence trimming.
  - Font Editor: support for editing graphics characters.